pyproject.toml
[project]
name = "nexus-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "py-cord>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
